Move czifile fix to a separate file
