+++ package/PartSegImage/image_reader.py
@@ -1,21 +1,17 @@
-import inspect
 import os.path
 import typing
 from abc import abstractmethod
 from contextlib import suppress
-from importlib.metadata import version
 from io import BytesIO
 from itertools import zip_longest
 from pathlib import Path
 from threading import Lock
 
-import imagecodecs
 import numpy as np
 import tifffile
-from czifile.czifile import DECOMPRESS, CziFile
+from czifile.czifile import CziFile
 from defusedxml import ElementTree
 from oiffile import OifFile
-from packaging.version import parse as parse_version
 
 from PartSegImage.image import ChannelInfo, Image
 
@@ -27,72 +23,6 @@ if typing.TYPE_CHECKING:
 
 
 CZI_MAX_WORKERS = None
-
-
-class ZSTD1Header(typing.NamedTuple):
-    """
-    ZSTD1 header structure
-    based on:
-    https://github.com/ZEISS/libczi/blob/4a60e22200cbf0c8ff2a59f69a81ef1b2b89bf4f/Src/libCZI/decoder_zstd.cpp#L19
-    """
-
-    header_size: int
-    hiLoByteUnpackPreprocessing: bool
-
-
-def parse_zstd1_header(data: bytes, size: int) -> ZSTD1Header:  # pragma: no cover
-    """
-    Parse ZSTD header
-
-    https://github.com/ZEISS/libczi/blob/4a60e22200cbf0c8ff2a59f69a81ef1b2b89bf4f/Src/libCZI/decoder_zstd.cpp#L84
-    """
-    if size < 1:
-        return ZSTD1Header(0, False)
-
-    if data[0] == 1:
-        return ZSTD1Header(1, False)
-
-    if data[0] == 3 and size < 3:
-        return ZSTD1Header(0, False)
-
-    if data[1] == 1:
-        return ZSTD1Header(3, bool(data[2] & 1))
-
-    return ZSTD1Header(0, False)
-
-
-def _get_dtype():
-    return inspect.currentframe().f_back.f_back.f_locals["de"].dtype
-
-
-def decode_zstd1(data: bytes) -> np.ndarray:
-    """
-    Decode ZSTD1 data
-    """
-    header = parse_zstd1_header(data, len(data))
-    dtype = _get_dtype()
-    if header.hiLoByteUnpackPreprocessing:
-        array_ = np.frombuffer(imagecodecs.zstd_decode(data[header.header_size :]), np.uint8).copy()
-        half_size = array_.size // 2
-        array = np.empty(half_size, np.uint16)
-        array[:] = array_[:half_size] + (array_[half_size:].astype(np.uint16) << 8)
-        array = array.view(dtype)
-    else:
-        array = np.frombuffer(imagecodecs.zstd_decode(data[header.header_size :]), dtype).copy()
-    return array
-
-
-def decode_zstd0(data: bytes) -> np.ndarray:
-    """
-    Decode ZSTD0 data
-    """
-    dtype = _get_dtype()
-    return np.frombuffer(imagecodecs.zstd_decode(data), dtype).copy()
-
-
-if parse_version(version("czifile")) in {parse_version("2019.7.2"), parse_version("2019.7.2.1")}:
-    DECOMPRESS[5] = decode_zstd0
-    DECOMPRESS[6] = decode_zstd1
 
 
 def _empty(_, __):
